+++ spectre/src/lib.rs
@@ -17,6 +17,10 @@
 //!     Ok(())
 //! }
 //! ```
+//!
+//! # Pages
+//!
+//! [`Page`]'s are individual chrome tabs that can be controlled.
 mod browser;
 pub mod cdp;
 pub mod dom;
@@ -30,3 +34,15 @@ pub use page::Page;
 pub use spectre_macros::test;
 
 pub const EMPTY_PAGE: &str = "https://blank.org/";
+
+/// Get any available port on the device
+pub async fn get_available_port() -> Result<u16> {
+    // Get any available port
+    let listener = std::net::TcpListener::bind("127.0.0.1:0")?;
+    let port = listener.local_addr()?.port();
+
+    // Immediately drop the listener to free the port
+    std::mem::drop(listener);
+
+    Ok(port)
+}