+++ spectre/tests/cdp_messages.rs
@@ -2,32 +2,30 @@
 //! responses are parsed correctly
 use serde_json::Value;
 use spectre::{
-    Browser, Result,
+    Result,
     cdp::{
-        AttachToTargetResponse, CDPConnection, CDPMessage, CDPMethod, CreateTargetResponse,
+        AttachToTargetResponse, CdpConnection, CDPMessage, CdpMethod, CreateTargetResponse,
         GetDocumentResponse, GetTargetResponse, PageNavigateResponse,
     },
     dom::NodeName,
 };
 
-#[tokio::test]
+#[spectre::test]
 async fn get_targets() -> Result<()> {
-    let browser = Browser::start().await?;
-    let conn = CDPConnection::new(browser.url()).await?;
-    let message = CDPMessage::root(1, CDPMethod::GetTargets);
+    let conn = CdpConnection::new(&browser.url()).await?;
+    let message = CDPMessage::root(1, CdpMethod::GetTargets);
     let response: GetTargetResponse = conn.send(message).await?;
     assert_eq!(response.id(), 1);
 
     Ok(())
 }
 
-#[tokio::test]
+#[spectre::test]
 async fn create_target() -> Result<()> {
-    let browser = Browser::start().await?;
-    let conn = CDPConnection::new(browser.url()).await?;
+    let conn = CdpConnection::new(&browser.url()).await?;
     let message = CDPMessage::root(
         1,
-        CDPMethod::CreateTarget {
+        CdpMethod::CreateTarget {
             url: String::from("https://example.com"),
         },
     );
@@ -36,16 +34,15 @@ async fn create_target() -> Result<()> {
     Ok(())
 }
 
-#[tokio::test]
+#[spectre::test]
 async fn attach_to_target() -> Result<()> {
-    let browser = Browser::start().await?;
-    let conn = CDPConnection::new(browser.url()).await?;
+    let conn = CdpConnection::new(&browser.url()).await?;
     let message = CDPMessage::get_targets(1);
 
     let response: GetTargetResponse = conn.send(message).await?;
     let targets = response.body().targets;
 
-    let method = CDPMethod::AttachToTarget {
+    let method = CdpMethod::AttachToTarget {
         target_id: targets[0].target_id.clone(),
         flatten: true,
     };
@@ -55,16 +52,15 @@ async fn attach_to_target() -> Result<()> {
     Ok(())
 }
 
-#[tokio::test]
+#[spectre::test]
 async fn page_navigate() -> Result<()> {
-    let browser = Browser::start().await?;
-    let conn = CDPConnection::new(browser.url()).await?;
+    let conn = CdpConnection::new(&browser.url()).await?;
     let message = CDPMessage::get_targets(1);
 
     let response: GetTargetResponse = conn.send(message).await?;
     let targets = response.body().targets;
 
-    let method = CDPMethod::AttachToTarget {
+    let method = CdpMethod::AttachToTarget {
         target_id: targets[0].target_id.clone(),
         flatten: true,
     };
@@ -78,16 +74,15 @@ async fn page_navigate() -> Result<()> {
     Ok(())
 }
 
-#[tokio::test]
+#[spectre::test]
 async fn get_document() -> Result<()> {
-    let mut browser = Browser::start().await?;
     let page = browser.goto("https://example.com").await?;
 
     let url = page.endpoint();
-    let conn = CDPConnection::new(url).await?;
+    let conn = CdpConnection::new(url).await?;
 
     // Set to -1 to get the whole tree
-    let method = CDPMethod::GetDocument { depth: -1 };
+    let method = CdpMethod::GetDocument { depth: -1 };
     let message = CDPMessage::root(2, method);
     let response: GetDocumentResponse = conn.send(message).await?;
 
@@ -96,16 +91,15 @@ async fn get_document() -> Result<()> {
     Ok(())
 }
 
-#[tokio::test]
+#[spectre::test]
 async fn page_navigate_error() -> Result<()> {
-    let browser = Browser::start().await?;
-    let conn = CDPConnection::new(browser.url()).await?;
+    let conn = CdpConnection::new(&browser.url()).await?;
     let message = CDPMessage::get_targets(1);
 
     let response: GetTargetResponse = conn.send(message).await?;
     let targets = response.body().targets;
 
-    let method = CDPMethod::AttachToTarget {
+    let method = CdpMethod::AttachToTarget {
         target_id: targets[0].target_id.clone(),
         flatten: true,
     };
@@ -120,15 +114,15 @@ async fn page_navigate_error() -> Result<()> {
     Ok(())
 }
 
-#[tokio::test]
+
+#[spectre::test]
 async fn runtime_evaluate() -> Result<()> {
-    let browser = Browser::start().await?;
-    let conn = CDPConnection::new(browser.url()).await?;
+    let conn = CdpConnection::new(&browser.url()).await?;
 
     let response: GetTargetResponse = conn.send(CDPMessage::get_targets(1)).await?;
     let targets = response.body().targets;
 
-    let method = CDPMethod::AttachToTarget {
+    let method = CdpMethod::AttachToTarget {
         target_id: targets[0].target_id.clone(),
         flatten: true,
     };
@@ -136,7 +130,7 @@ async fn runtime_evaluate() -> Result<()> {
     let response: AttachToTargetResponse = conn.send(message).await?;
     let session_id = response.body().session_id;
 
-    let method = CDPMethod::Evaluate {
+    let method = CdpMethod::Evaluate {
         expression: String::from(
             "
 			function hi(){
@@ -152,6 +146,34 @@ async fn runtime_evaluate() -> Result<()> {
     let message = CDPMessage::new(1, &session_id, method);
     let response: Value = conn.send(message).await?;
     dbg!(&response);
+
+    Ok(())
+}
+
+#[spectre::test]
+async fn send_cdp_message() -> Result<()> {
+    let ws_url = browser.url();
+
+    let conn = CdpConnection::new(&ws_url).await?;
+    let message = CDPMessage::root(2, CdpMethod::GetTargets);
+    let _: GetTargetResponse = conn.send(message).await?;
+
+    Ok(())
+}
+
+#[spectre::test]
+async fn multiple_connections() -> Result<()> {
+    let ws_url = browser.url();
+
+    let conn1 = CdpConnection::new(&ws_url).await?;
+    let conn2 = CdpConnection::new(&ws_url).await?;
+
+    let _: GetTargetResponse = conn1
+        .send(CDPMessage::root(2, CdpMethod::GetTargets))
+        .await?;
+    let _: GetTargetResponse = conn2
+        .send(CDPMessage::root(2, CdpMethod::GetTargets))
+        .await?;
 
     Ok(())
 }