+++ spectre/tests/cdp_session.rs
@@ -1,23 +1,21 @@
 use serde_json::json;
 use spectre::{
-    Browser, EMPTY_PAGE, Error, Result,
-    cdp::{CDPConnection, CDPMethod, GetTargetResponse},
+    EMPTY_PAGE, Error, Result,
+    cdp::{CdpConnection, CdpMethod, GetTargetResponse},
 };
 
-#[tokio::test]
+#[spectre::test]
 async fn create_session() -> Result<()> {
-    let browser = Browser::start().await?;
-    let connection = CDPConnection::new(browser.url()).await?;
+    let connection = CdpConnection::new(&browser.url()).await?;
     let session = connection.create_session().await?;
 
-    let _: GetTargetResponse = session.send(CDPMethod::GetTargets).await?;
+    let _: GetTargetResponse = session.send(CdpMethod::GetTargets).await?;
     Ok(())
 }
 
-#[tokio::test]
+#[spectre::test]
 async fn get_box_model() -> Result<()> {
-    let browser = Browser::start().await?;
-    let connection = CDPConnection::new(browser.url()).await?;
+    let connection = CdpConnection::new(&browser.url()).await?;
     let session = connection.create_session().await?;
 
     let root = session.get_dom(-1).await?;
@@ -26,30 +24,27 @@ async fn get_box_model() -> Result<()> {
     Ok(())
 }
 
-#[tokio::test]
+#[spectre::test]
 async fn navigate() -> Result<()> {
-    let browser = Browser::start().await?;
-    let connection = CDPConnection::new(browser.url()).await?;
+    let connection = CdpConnection::new(&browser.url()).await?;
     let session = connection.create_session().await?;
 
     session.navigate(EMPTY_PAGE).await?;
     Ok(())
 }
 
-#[tokio::test]
+#[spectre::test]
 async fn evaluate() -> Result<()> {
-    let browser = Browser::start().await?;
-    let connection = CDPConnection::new(browser.url()).await?;
+    let connection = CdpConnection::new(&browser.url()).await?;
     let session = connection.create_session().await?;
 
     let _ = session.evaluate("5").await?;
     Ok(())
 }
 
-#[tokio::test]
+#[spectre::test]
 async fn can_handle_exception() -> Result<()> {
-    let browser = Browser::start().await?;
-    let connection = CDPConnection::new(browser.url()).await?;
+    let connection = CdpConnection::new(&browser.url()).await?;
     let session = connection.create_session().await?;
 
     let result = session.evaluate("throw 5").await;
@@ -69,10 +64,9 @@ async fn can_handle_exception() -> Result<()> {
     Ok(())
 }
 
-#[tokio::test]
+#[spectre::test]
 async fn can_handle_syntax_error() -> Result<()> {
-    let browser = Browser::start().await?;
-    let connection = CDPConnection::new(browser.url()).await?;
+    let connection = CdpConnection::new(&browser.url()).await?;
     let session = connection.create_session().await?;
 
     let result = session.evaluate("return 5").await;
@@ -80,10 +74,9 @@ async fn can_handle_syntax_error() -> Result<()> {
     Ok(())
 }
 
-#[tokio::test]
+#[spectre::test]
 async fn query_selector_handle_missing_element() -> Result<()> {
-    let browser = Browser::start().await?;
-    let connection = CDPConnection::new(browser.url()).await?;
+    let connection = CdpConnection::new(&browser.url()).await?;
     let session = connection.create_session().await?;
 
     let root = session.get_dom(-1).await?;