+++ CHANGELOG.md
@@ -5,6 +5,10 @@ commit messages/pull requests but rather a hand curated list of notable changes.
 
 ## [Unreleased]
 
+### Added
+
+- Added `spectre::test` attribute macro, this macro reuses the same browser instance for every test.
+
 ### Changed
 
 - Download chrome to `$HOME/.spectre/browsers` instead of the the current working directory.
@@ -17,5 +21,5 @@ commit messages/pull requests but rather a hand curated list of notable changes.
 - Added `Browser` struct which starts a chrome child process for communication.
 - Added `Page` struct for controlling a single browser tab.
 - Added `spectre::test` macro which uses the same browser between tests (experimental).
-- Added `CDPConnection` and `CDPSession` for sending and receiving messages to the browser via the Chrome DevTools protocol.
+- Added `CdpConnection` and `CdpSession` for sending and receiving messages to the browser via the Chrome DevTools protocol.